Fix 24 bpp / 3 Bpp handling.
draw_tile3*() functions now use the same algorithm as the others except
for the heavy use of memcpy() in order to copy 3 bytes at once. Slow but
no one cares about 24 bpp anyway. The reason for this change is that the
old algorithm swapped color channels.

This commit also implements the missing scaling algorithm for this depth
in sdl.cpp.

Signed-off-by: zamaz <zamaz@users.sourceforge.net>