Remove gst_name()
Two issues with this function, the way a path is stripped from its
directory doesn't work on all systems and it returns a pointer to a
local buffer.

Rename md::romfilename to md::romname, remove gst_name(), add
dgen_basename() and use it directly in md() to strip the file name from
its directory and extension to store the result in md::romname which
doesn't need the entire path.

Signed-off-by: zamaz <zamaz@users.sourceforge.net>