sdl: implement prompt
This prompt provides a way to change any rc variable at runtime and
comes with history management and command completion.

It's displayed in the message bar and can be triggered by pressing ":"
(this is the default key). Just like the Game Genie prompt, it is
persistent when in stopped mode.

Signed-off-by: zamaz <zamaz@users.sourceforge.net>