sdl: add scale2x support
Scale2x is much faster than hqx but less accurate. It has been added to
the list of scaling filters, is enabled by default by configure but can
be disabled with --disable-scale2x.

It's available in DGen for 15, 16 and 32 bpp, and supports scaling
factors of 2x2, 2x3, 2x4, 3x3 and 3x4.

Signed-off-by: zamaz <zamaz@users.sourceforge.net>