vdp: allow dgen to deal with half-submitted commands
VDP commands are two 16-bit writes into the VDP data port. When the
first word arrives, the top 14 bits of the address and the two CD bits
are immediately updated in the VDP state. Then when the second word
arrives the remaining address and CD bits are updated. Before, dgen
would cache the first word and wait for the second, committing changes
when the second word arrived; this is wrong. Further, a read from the
data port or a register should abort a pending command.

This commit finally fixes the mangled EA logo.

Acked-by: zamaz <zamaz@users.sourceforge.net>
Signed-off-by: egelmex <ellism88@gmail.com>
Signed-off-by: vext01 <vext01@gmail.com>