Enhance region selection settings
This commit enables DGen/SDL to update video, sound and other settings
when configuring region, either manually from the command-line, or from
the "region" RC variable.

When unspecified (now the default), it is derived from the ROM header.

Also added the 'X' pseudo region, which is Japan (J) in PAL mode.

Signed-off-by: zamaz <zamaz@users.sourceforge.net>