sdl: implement optional thread for displaying frames
When bool_screen_thread is enabled, DGen/SDL creates a separate thread
to offload the displaying of frames.

This is only useful on slower machines where flipping video buffers
takes time, especially when V-sync is enabled and doing so blocks
DGen/SDL until the next frame without consuming CPU time (such as the
case of Dispmanx on the Raspberry Pi when bool_doublebuffer is also
enabled).

This requires threads to be compiled-in (--enable-threads) and doesn't
work with OpenGL (apparently an SDL issue).

Signed-off-by: zamaz <zamaz@users.sourceforge.net>