ras: improve sprites priority bit handling
When they overlap, sprites with the high priority bit set can be
displayed behind sprites without it *if* the latter have a lower index
in the sprites list. In that case, the priority bit must be ignored.

This fixes games such as Streets of Rage 1 in which items and weapons
always appear on top of other sprites.

Signed-off-by: zamaz <zamaz@users.sourceforge.net>