ras: fix sprites priority bit handling
The previous commit regarding this feature managed it on a per-sprite
basis, by rendering sprites of the high priority layer on the low
priority layer when necessary.

The VDP doesn't work like that. Sprites aren't supposed to move between
layers, and the overlapping is managed on a per-pixel basis.

This commit reimplements all of this properly.

Signed-off-by: zamaz <zamaz@users.sourceforge.net>