sdl: reimplement calibration command
This new implementation is similar to the last one (which was
disabled until now), as it uses a CTV filter that displays some ugly
text asking to press various buttons, but it now accepts any key or
joystick axis/button to do so.

Because it's not only for the joystick anymore, "calibrate_js" has been
renamed "calibrate" (both commands are actually aliased).

Signed-off-by: zamaz <zamaz@users.sourceforge.net>