debug: fix instructions counters
Since the previous commit, instructions counters are never updated when
no breakpoints or watchpoints are set. This was done for performance
reasons to execute more than a single instruction during each iteration.

This commit adds a new command "count" to toggle these counters and
force a single instruction per iteration.

Signed-off-by: zamaz <zamaz@users.sourceforge.net>