Support key/joystick button combinations bindings
An annoying problem when controlling DGen/SDL from a joystick/joypad is
that the number of buttons is limited, as each button can only perform
a single action. Binding commands (such as "exit") to a single button
press is also dangerous.

This commit adds support for combinations by extending the bind_* RC
syntax.

Signed-off-by: zamaz <zamaz@users.sourceforge.net>